Modularize DataTable more
