+++ packages/design-system/src/components/DataTable/DataTable.utils.ts
@@ -6,7 +6,19 @@
 // SPDX-License-Identifier: MIT
 //
 import { type RankingInfo, rankItem } from '@tanstack/match-sorter-utils'
-import type { FilterFn } from '@tanstack/react-table'
+import {
+  FilterFn,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
+  SortingState,
+  useReactTable,
+} from '@tanstack/react-table'
+import { useState } from 'react'
+import { useDebouncedCallback } from 'use-debounce'
+import type { PartialSome } from '@/packages/design-system/src/utils/misc'
+import type { TableOptions } from '@tanstack/table-core'
 
 declare module '@tanstack/react-table' {
   interface FilterFns {
@@ -33,4 +45,47 @@ export const fuzzyFilter: FilterFn<unknown> = (
 
   // Return if the item should be filtered in/out
   return itemRank.passed
+}
+
+export interface UseDataTableProps<Data>
+  extends PartialSome<TableOptions<Data>, 'getCoreRowModel' | 'filterFns'> {
+  pageSize?: number
+}
+
+export const useDataTable = <Data>({
+  columns,
+  data,
+  pageSize = 50,
+  ...props
+}: UseDataTableProps<Data>) => {
+  const [sorting, setSorting] = useState<SortingState>([])
+  const [globalFilter, setGlobalFilter] = useState('')
+  const setGlobalFilterDebounced = useDebouncedCallback(
+    (value: string) => setGlobalFilter(value),
+    200,
+  )
+
+  const table = useReactTable({
+    columns,
+    data,
+    filterFns: {
+      fuzzy: fuzzyFilter,
+    },
+    globalFilterFn: 'fuzzy',
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    onGlobalFilterChange: setGlobalFilter,
+    onSortingChange: setSorting,
+    state: { globalFilter, sorting },
+    initialState: {
+      pagination: {
+        pageSize,
+      },
+    },
+    ...props,
+  })
+
+  return { globalFilter, setGlobalFilter, setGlobalFilterDebounced, table }
 }
