+++ packages/design-system/src/components/DataTable/index.tsx
@@ -8,3 +8,4 @@
 export * from './DataTable'
 export * from './RowDropdownMenu'
 export * from './DataTable.columns'
+export * from './DataTable.utils'
